+++ sdk/monitor/azure-monitor-opentelemetry-exporter/tests/statsbeat/test_statsbeat.py
@@ -29,6 +29,7 @@ from azure.monitor.opentelemetry.exporter.statsbeat._state import (
 from azure.monitor.opentelemetry.exporter.statsbeat._statsbeat_metrics import (
     _shorten_host,
     _FEATURE_TYPES,
+    _StatsbeatFeature,
     _StatsbeatMetrics,
     _RP_NAMES,
 )
@@ -53,9 +54,15 @@ class TestStatsbeat(unittest.TestCase):
     def setUp(self):
         _statsbeat._STATSBEAT_METER_PROVIDER = None
 
+    @mock.patch.object(MeterProvider, 'shutdown')
+    @mock.patch.object(MeterProvider, 'force_flush')
     @mock.patch.object(_StatsbeatMetrics, 'init_non_initial_metrics')
-    @mock.patch.object(_StatsbeatMetrics, '_get_attach_metric')
-    def test_collect_statsbeat_metrics(self, attach_mock, non_init_mock):
+    def test_collect_statsbeat_metrics(
+        self,
+        non_init_mock,
+        flush_mock,
+        shutdown_mock
+    ):
         exporter = mock.Mock()
         exporter._endpoint = "test endpoint"
         exporter._instrumentation_key = "test ikey"
@@ -68,8 +75,8 @@ class TestStatsbeat(unittest.TestCase):
         self.assertTrue(isinstance(mr, PeriodicExportingMetricReader))
         self.assertIsNotNone(mr._exporter)
         self.assertTrue(isinstance(mr._exporter, _StatsBeatExporter))
-        attach_mock.assert_called_once()
         non_init_mock.assert_called_once()
+        flush_mock.assert_called_once()
 
     def test_collect_statsbeat_metrics_exists(self):
         exporter = mock.Mock()
@@ -79,9 +86,15 @@ class TestStatsbeat(unittest.TestCase):
         _statsbeat.collect_statsbeat_metrics(exporter)
         self.assertEqual(_statsbeat._STATSBEAT_METER_PROVIDER, mock_mp)
 
+    @mock.patch.object(MeterProvider, 'shutdown')
+    @mock.patch.object(MeterProvider, 'force_flush')
     @mock.patch.object(_StatsbeatMetrics, 'init_non_initial_metrics')
-    @mock.patch.object(_StatsbeatMetrics, '_get_attach_metric')
-    def test_collect_statsbeat_metrics_non_eu(self,  attach_mock, non_init_mock):
+    def test_collect_statsbeat_metrics_non_eu(
+        self,
+        non_init_mock,
+        flush_mock,
+        shutdown_mock
+    ):
         exporter = mock.Mock()
         exporter._instrumentation_key = "1aa11111-bbbb-1ccc-8ddd-eeeeffff3333"
         exporter._endpoint = "https://westus-0.in.applicationinsights.azure.com/"
@@ -89,9 +102,7 @@ class TestStatsbeat(unittest.TestCase):
         with mock.patch.dict(
             os.environ, {
                 "APPLICATION_INSIGHTS_STATS_CONNECTION_STRING": "",
-            }), \
-            mock.patch.object(_StatsbeatMetrics, 'init_non_initial_metrics'), \
-            mock.patch.object(_StatsbeatMetrics, '_get_attach_metric'):
+            }):
             _statsbeat.collect_statsbeat_metrics(exporter)
         mp = _statsbeat._STATSBEAT_METER_PROVIDER
         mr = mp._sdk_config.metric_readers[0]
@@ -105,7 +116,15 @@ class TestStatsbeat(unittest.TestCase):
             _statsbeat._DEFAULT_NON_EU_STATS_CONNECTION_STRING.split(";")[1].split("=")[1]   # noqa: E501
         )
 
-    def test_collect_statsbeat_metrics_eu(self):
+    @mock.patch.object(MeterProvider, 'shutdown')
+    @mock.patch.object(MeterProvider, 'force_flush')
+    @mock.patch.object(_StatsbeatMetrics, 'init_non_initial_metrics')
+    def test_collect_statsbeat_metrics_eu(
+        self,
+        non_init_mock,
+        flush_mock,
+        shutdown_mock
+    ):
         exporter = mock.Mock()
         exporter._instrumentation_key = "1aa11111-bbbb-1ccc-8ddd-eeeeffff3333"
         exporter._endpoint = "https://northeurope-0.in.applicationinsights.azure.com/"
@@ -113,9 +132,7 @@ class TestStatsbeat(unittest.TestCase):
         with mock.patch.dict(
             os.environ, {
                 "APPLICATION_INSIGHTS_STATS_CONNECTION_STRING": "",
-            }), \
-            mock.patch.object(_StatsbeatMetrics, 'init_non_initial_metrics'), \
-            mock.patch.object(_StatsbeatMetrics, '_get_attach_metric'):
+            }):
             _statsbeat.collect_statsbeat_metrics(exporter)
         mp = _statsbeat._STATSBEAT_METER_PROVIDER
         mr = mp._sdk_config.metric_readers[0]
@@ -135,6 +152,13 @@ _StatsbeatMetrics_COMMON_ATTRS = dict(
 _StatsbeatMetrics_NETWORK_ATTRS = dict(
     _StatsbeatMetrics._NETWORK_ATTRIBUTES
 )
+_StatsbeatMetrics_FEATURE_ATTRIBUTES = dict(
+    _StatsbeatMetrics._FEATURE_ATTRIBUTES
+)
+_StatsbeatMetrics_INSTRUMENTATION_ATTRIBUTES = dict(
+    _StatsbeatMetrics._INSTRUMENTATION_ATTRIBUTES
+)
+
 
 # pylint: disable=protected-access
 class TestStatsbeatMetrics(unittest.TestCase):
@@ -159,6 +183,12 @@ class TestStatsbeatMetrics(unittest.TestCase):
         )
         _StatsbeatMetrics._NETWORK_ATTRIBUTES = dict(
             _StatsbeatMetrics_NETWORK_ATTRS
+        )
+        _StatsbeatMetrics._FEATURE_ATTRIBUTES = dict(
+            _StatsbeatMetrics_FEATURE_ATTRIBUTES
+        )
+        _StatsbeatMetrics._INSTRUMENTATION_ATTRIBUTES = dict(
+            _StatsbeatMetrics_INSTRUMENTATION_ATTRIBUTES
         )
         _REQUESTS_MAP.clear()
 
@@ -481,14 +511,14 @@ class TestStatsbeatMetrics(unittest.TestCase):
         attributes = dict(_StatsbeatMetrics._COMMON_ATTRIBUTES)
         attributes.update(_StatsbeatMetrics._FEATURE_ATTRIBUTES)
         self.assertEqual(attributes["feature"], 1)
-        self.assertEqual(attributes["type"], 0)
-        observations = self._metric._get_feature_metric(options=None)
+        self.assertEqual(attributes["type"], _FEATURE_TYPES.FEATURE)
+        observations = metric._get_feature_metric(options=None)
         for obs in observations:
             self.assertEqual(obs.value, 1)
             self.assertEqual(obs.attributes, attributes)
 
     # pylint: disable=protected-access
-    def test_get_feature_metric_no_local_storage(self):
+    def test_get_feature_metric_none(self):
         mp = MeterProvider()
         ikey = "1aa11111-bbbb-1ccc-8ddd-eeeeffff3334"
         endpoint = "https://westus-1.in.applicationinsights.azure.com/"
@@ -499,14 +529,64 @@ class TestStatsbeatMetrics(unittest.TestCase):
             False,
             0,
         )
+        self.assertEqual(_StatsbeatMetrics._FEATURE_ATTRIBUTES["feature"], 0)
+        observations = metric._get_feature_metric(options=None)
+        self.assertEqual(len(observations), 0)
+        self.assertEqual(_StatsbeatMetrics._FEATURE_ATTRIBUTES["feature"], 0)
+
+    # pylint: disable=protected-access
+    def test_get_feature_metric_instrumentation(self):
+        mp = MeterProvider()
+        ikey = "1aa11111-bbbb-1ccc-8ddd-eeeeffff3334"
+        endpoint = "https://westus-1.in.applicationinsights.azure.com/"
+        metric = _StatsbeatMetrics(
+            mp,
+            ikey,
+            endpoint,
+            False,
+            0,
+        )
+        metric._feature = _StatsbeatFeature.NONE
         attributes = dict(_StatsbeatMetrics._COMMON_ATTRIBUTES)
-        attributes.update(_StatsbeatMetrics._FEATURE_ATTRIBUTES)
+        attributes.update(_StatsbeatMetrics._INSTRUMENTATION_ATTRIBUTES)
+        self.assertEqual(attributes["type"], _FEATURE_TYPES.INSTRUMENTATION)
         self.assertEqual(attributes["feature"], 0)
-        self.assertEqual(attributes["type"], 0)
-        observations = self._metric._get_feature_metric(options=None)
+        with mock.patch(
+            "azure.monitor.opentelemetry.exporter._utils.get_instrumentations"
+        ) as instrumentations:
+            instrumentations.return_value = 1026
+            observations = metric._get_feature_metric(options=None)
+        self.assertEqual(
+            _StatsbeatMetrics._INSTRUMENTATION_ATTRIBUTES["feature"],
+            1026,
+        )
+        attributes["feature"] = 1026
         for obs in observations:
             self.assertEqual(obs.value, 1)
             self.assertEqual(obs.attributes, attributes)
+        self.assertEqual(_StatsbeatMetrics._INSTRUMENTATION_ATTRIBUTES["feature"], 1026)
+
+    # pylint: disable=protected-access
+    def test_get_feature_metric_instrumentation_none(self):
+        mp = MeterProvider()
+        ikey = "1aa11111-bbbb-1ccc-8ddd-eeeeffff3334"
+        endpoint = "https://westus-1.in.applicationinsights.azure.com/"
+        metric = _StatsbeatMetrics(
+            mp,
+            ikey,
+            endpoint,
+            False,
+            0,
+        )
+        metric._feature = _StatsbeatFeature.NONE
+        self.assertEqual(_StatsbeatMetrics._INSTRUMENTATION_ATTRIBUTES["feature"], 0)
+        with mock.patch(
+            "azure.monitor.opentelemetry.exporter._utils.get_instrumentations"
+        ) as instrumentations:
+            instrumentations.return_value = 0
+            observations = metric._get_feature_metric(options=None)
+        self.assertEqual(len(observations), 0)
+        self.assertEqual(_StatsbeatMetrics._INSTRUMENTATION_ATTRIBUTES["feature"], 0)
 
     def test_init_non_initial_metrics(self):
         mp = MeterProvider()